V1.1 Update
Changed interface to represent a softphone style.
Launching without a callto link will allow to input a number manually and dial a number that way.
Note, it is not possible to actually send additional characters, hence no # is present on the keypad.

Settings are now placed into a separate settings button.