Changed API Request from using a webbrowser element to a webrequest
Previously, a webbrowser element was navigated to the api page of the yealink phones to make the call request.

This seems to cause issues sometimes. (https://github.com/DJFPaul/yealink_dialer/issues/2)
Additionally this is not the most secure.

This changes the way requests are made from using a webbrowser, to a HTTPWebRequest