Bugfixes and feature changes
Fixed the testing / save button working properly now.

Before it already reconfigured the active variable when pressing the test button
.
While with current ui this would not have caused a issue as you can't exit the settings without saving, it would have if a button to exit without saving was ever introduced.
It now is fixed to make it clear at what stage what happens and prevent future issue potential.